Eject from shared workflows (#2)
The template repository from which this repository was generated uses
shared workflows[^1] on GitHub. This approach had the advantage of
quickly bootstrapping new repositories, but introduced challenges later
on when workflows needed to be customized. This repository is actually
part of a larger effort to create a better solution to this problem.

[^1]: https://docs.github.com/en/actions/using-workflows/reusing-workflows